requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
faiss-cpu==1.8.0.post1
fastapi==0.112.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Levenshtein==0.25.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.3
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
rapidfuzz==3.9.5
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.37.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
watchdog==4.0.1
Whoosh==2.7.4
